Finished Beta Stable Release
"That's fine, it's a feature, not a bug." Did the work without github at the beginning due to technical difficulties. This is the first release of "Synthwave Driver" A relaxing drive around game with no real goal but to drive, vibe, and relax. By all means download and try out the game itself kept in the folder "SynthwaveDriver1.1"